+++ example/lib/main.dart
@@ -36,34 +36,18 @@ class _HomeViewState extends State<HomeView> {
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(title: const Text("Home")),
-        body: Center(
-            child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            if (selectedCountry != null)
-              Row(mainAxisSize: MainAxisSize.min, children: [
-                const SizedBox(width: 11),
-                //INCLUDE THE PACKAGE NAME  IN Image.asset()
-                Image.asset(selectedCountry?.icon ?? "",
-                    height: 21, width: 21, package: "dialcodeselector"),
-                const SizedBox(width: 7),
-                Padding(
-                  padding: const EdgeInsets.only(right: 13),
-                  child: Text(selectedCountry?.dialCode ?? "",
-                      style: const TextStyle(fontSize: 16)),
-                )
-              ]),
-            TextButton(
-                child: const Text("Choose Country"),
-                onPressed: () async {
-                  Country? value = await DialCodeSelector.selectCountry(context,
-                      initialCountryCode: initialCountrCode);
-                  if (value == null) return;
-                  selectedCountry = value;
-                  initialCountrCode = selectedCountry?.dialCode ?? "";
-                  setState(() {});
-                })
-          ],
-        )));
+        body: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 25),
+          child: Center(
+              child: TextFormField(
+            decoration: InputDecoration(
+                hintText: "Enter phone number",
+                prefixIcon: DialCodeSelector(callback: (selectedCountry) {
+                  //use selectedCountry from callback here, Country logo is Null here
+                  print("Selected Country: ${selectedCountry.countryName}");
+                })),
+            onChanged: (String value) {},
+          )),
+        ));
   }
 }