+++ lib/dialcodeselector.dart
@@ -1,4 +1,5 @@
 library dialcodeselector;
 
-export 'src/Main/diacodeSelector.dart';
+export 'src/Main/dialCodeSelectorView.dart';
 export 'src/Model/country.dart';
+export 'src/Model/dialCodeSelectorTheme.dart';