+++ lib/src/Model/dialCodeSelectorTheme.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class DialCodeSelectorTheme {
+  final Color? dialCodeColor;
+  final Color? dropDownButtonColor;
+  final Color? selectedDialCodeColor;
+  final Color? countryNameColor;
+  DialCodeSelectorTheme(
+      {this.dialCodeColor,
+      this.countryNameColor,
+      this.dropDownButtonColor,
+      this.selectedDialCodeColor});
+}